Move PVF workers into separate crate
