+++ node/core/pvf/Cargo.toml
@@ -4,32 +4,15 @@ version.workspace = true
 authors.workspace = true
 edition.workspace = true
 
-[[bin]]
-name = "prepare_worker"
-path = "bin/prepare_worker.rs"
-
-[[bin]]
-name = "execute_worker"
-path = "bin/execute_worker.rs"
-
-[[bin]]
-name = "puppet_worker"
-path = "bin/puppet_worker.rs"
-
 [dependencies]
 always-assert = "0.1"
-assert_matches = "1.4.0"
-cpu-time = "1.0.0"
 futures = "0.3.21"
 futures-timer = "3.0.2"
 gum = { package = "tracing-gum", path = "../../gum" }
 libc = "0.2.139"
 pin-project = "1.0.9"
 rand = "0.8.5"
-rayon = "1.5.1"
 slotmap = "1.0"
-tempfile = "3.3.0"
-tikv-jemalloc-ctl = { version = "0.5.0", optional = true }
 tokio = { version = "1.24.2", features = ["fs", "process"] }
 
 parity-scale-codec = { version = "3.4.0", default-features = false, features = ["derive"] }
@@ -38,13 +21,8 @@ polkadot-parachain = { path = "../../../parachain" }
 polkadot-core-primitives = { path = "../../../core-primitives" }
 polkadot-node-metrics = { path = "../../metrics" }
 polkadot-node-primitives = { path = "../../primitives" }
-
 polkadot-primitives = { path = "../../../primitives" }
-sc-executor = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-executor-wasmtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-executor-common = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-externalities = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "master" }
+
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-wasm-interface = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-maybe-compressed-blob = { git = "https://github.com/paritytech/substrate", branch = "master" }
@@ -53,14 +31,5 @@ sp-tracing = { git = "https://github.com/paritytech/substrate", branch = "master
 [build-dependencies]
 substrate-build-script-utils = { git = "https://github.com/paritytech/substrate", branch = "master" }
 
-[target.'cfg(target_os = "linux")'.dependencies]
-tikv-jemalloc-ctl = "0.5.0"
-
 [dev-dependencies]
-adder = { package = "test-parachain-adder", path = "../../../parachain/test-parachains/adder" }
-halt = { package = "test-parachain-halt", path = "../../../parachain/test-parachains/halt" }
 hex-literal = "0.3.4"
-tempfile = "3.3.0"
-
-[features]
-jemalloc-allocator = ["dep:tikv-jemalloc-ctl"]
