+++ node/core/pvf/src/lib.rs
@@ -29,11 +29,11 @@
 //!
 //! Then using the handle the client can send three types of requests:
 //!
-//! (a) PVF pre-checking. This takes the PVF [code][`Pvf`] and tries to prepare it (verify and
+//! (a) PVF pre-checking. This takes the `Pvf` code and tries to prepare it (verify and
 //! compile) in order to pre-check its validity.
 //!
 //! (b) PVF execution. This accepts the PVF [`params`][`polkadot_parachain::primitives::ValidationParams`]
-//!     and the PVF [code][`Pvf`], prepares (verifies and compiles) the code, and then executes PVF
+//!     and the `Pvf` code, prepares (verifies and compiles) the code, and then executes PVF
 //!     with the `params`.
 //!
 //! (c) Heads up. This request allows to signal that the given PVF may be needed soon and that it
@@ -93,7 +93,6 @@
 mod artifacts;
 mod error;
 mod execute;
-mod executor_intf;
 mod host;
 mod metrics;
 mod prepare;
@@ -101,64 +100,22 @@ mod priority;
 mod pvf;
 mod worker_common;
 
-#[doc(hidden)]
-pub mod testing;
-
-#[doc(hidden)]
-pub use sp_tracing;
-
+pub use artifacts::CompiledArtifact;
 pub use error::{InvalidCandidate, PrepareError, PrepareResult, ValidationError};
-pub use prepare::PrepareStats;
+pub use execute::{ExecuteHandshake, ExecuteResponse};
+#[cfg(any(target_os = "linux", feature = "jemalloc-allocator"))]
+pub use prepare::MemoryAllocationStats;
+pub use prepare::{MemoryStats, PrepareStats};
 pub use priority::Priority;
 pub use pvf::PvfPrepData;
 
 pub use host::{start, Config, ValidationHost};
 pub use metrics::Metrics;
-pub use worker_common::JOB_TIMEOUT_WALL_CLOCK_FACTOR;
-
-pub use execute::worker_entrypoint as execute_worker_entrypoint;
-pub use prepare::worker_entrypoint as prepare_worker_entrypoint;
-
-pub use executor_intf::{prepare, prevalidate};
-
-pub use sc_executor_common;
-pub use sp_maybe_compressed_blob;
+pub use worker_common::{framed_recv, framed_send, JOB_TIMEOUT_WALL_CLOCK_FACTOR};
 
 const LOG_TARGET: &str = "parachain::pvf";
 
-/// Use this macro to declare a `fn main() {}` that will create an executable that can be used for
-/// spawning the desired worker.
-#[macro_export(local_inner_macros)]
-macro_rules! decl_worker_main {
-	($command:tt) => {
-		fn main() {
-			$crate::sp_tracing::try_init_simple();
-
-			let args = std::env::args().collect::<Vec<_>>();
-
-			let mut version = None;
-			let mut socket_path: &str = "";
-
-			for i in 1..args.len() {
-				match args[i].as_ref() {
-					"--socket-path" => socket_path = args[i + 1].as_str(),
-					"--node-version" => version = Some(args[i + 1].as_str()),
-					_ => (),
-				}
-			}
-
-			decl_worker_main_command!($command, socket_path, version)
-		}
-	};
-}
-
-#[macro_export]
 #[doc(hidden)]
-macro_rules! decl_worker_main_command {
-	(prepare, $socket_path:expr, $version: expr) => {
-		$crate::prepare_worker_entrypoint(&$socket_path, $version)
-	};
-	(execute, $socket_path:expr, $version: expr) => {
-		$crate::execute_worker_entrypoint(&$socket_path, $version)
-	};
+pub mod testing {
+	pub use crate::worker_common::{spawn_with_program_path, SpawnErr};
 }
