+++ node/core/pvf/worker/build.rs
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Execute worker.
-// TODO: Build with musl.
-// TODO: Embed into polkadot binary.
-
-polkadot_node_core_pvf::decl_worker_main!(execute);
+fn main() {
+	substrate_build_script_utils::generate_cargo_keys();
+}
